+++ CProject/TESTMODEL/Src/network.cpp
@@ -12,7 +12,7 @@
 Matrix<85, 3> Conv_weight_ram;
 Matrix<20, 112, int8_t> fc2_weight_ram;
 
-void very_ugly_Linear(const int16_t *A, const int8_t *B, const int32_t *bias, const int32_t *right_shift, void (*act)(int32_t, int32_t, void *), uint32_t size_of_datatype, void *out, const int32_t M1_2, const int32_t M2_1, const int32_t M2_2);
+void very_ugly_Linear(const int16_t *A, const int8_t *B, const int32_t *bias, const int8_t *right_shift, void (*act)(int32_t, int8_t, void *), uint32_t size_of_datatype, void *out, const int32_t M1_2, const int32_t M2_1, const int32_t M2_2);
 
 void aiRun(const float input[1][1250][1], float result[2])
 {
@@ -92,7 +92,7 @@ void Model_Init()
 	// }	
 }
 
-void very_ugly_Linear(const int16_t *A, const int8_t *B, const int32_t *bias, const int32_t *right_shift, void (*act)(int32_t, int32_t, void *), uint32_t size_of_datatype, void *out, const int32_t M1_2, const int32_t M2_1, const int32_t M2_2)
+void very_ugly_Linear(const int16_t *A, const int8_t *B, const int32_t *bias, const int8_t *right_shift, void (*act)(int32_t, int32_t, void *), uint32_t size_of_datatype, void *out, const int32_t M1_2, const int32_t M2_1, const int32_t M2_2)
 {
 #pragma nounroll
 	for (int i2_1 = 0; i2_1 < M2_1; i2_1++)
